fix(esp32c3): real UART register map — Serial.print >128 bytes no longer wedges
The stats lab froze on "FETCHING STATS" with the panel never repainting. WiFi
was not the problem: the device associates, gets 192.168.4.2 and the HTTP body
arrives. The serial log stops at exactly 128 bytes into the body — the C3's
hardware TX FIFO depth.

configs/chips/esp32c3.yaml declared uart0/uart1 as the vendor-neutral
`type: "uart"`, which resolves to the STM32 register map (SR@0x00, DR@0x04).
On the C3 those offsets are FIFO@0x00 and STATUS@0x1C, so FIFO writes landed
on a status register, STATUS read back 0 (uart_ll_get_txfifo_len() saw all 128
entries free, always), and nothing ever asserted the UART interrupt.

ESP-IDF's blocking uart_tx_all() fills the TX FIFO and, for any write longer
than it, sets tx_waiting_fifo, enables UART_INTR_TXFIFO_EMPTY and blocks on
tx_fifo_sem. With no TX interrupt that semaphore was never given, so any
Serial.print over 128 bytes blocked forever: every task stalled and the idle
task parked in wfi (confirmed by a PC histogram — 20000/20000 samples on one
address, no interrupt for 4e9 cycles). Short lines fit one FIFO fill and hid
it; BROM ets_printf writes the FIFO directly and never waits on the driver,
which is why stray ROM log lines still appeared after the hang.

Both C3 UARTs now bind the real Espressif twin. The S3 already modeled this IP
faithfully (128-entry FIFOs, STATUS occupancy, the uart_ll.h interrupt set), and
the C3 carries the identical block, so esp32s3/uart.rs moves to the chip-neutral
peripherals/esp_uart.rs and both families wire it — keeping the arch-isolation
gate honest instead of letting one chip reach into another's model. Per-family
bits stay constructor arguments: matrix sources (C3 UART0=21/UART1=22) and the
core clock, which the twin's baud pacing needs at the C3's 160 MHz.

Also:
- EspUart gains rx_buffer(), and attach_uart_rx_source[_named] resolve it, so
  `uart_injections:` and interactive serial keep working on the C3.
- attach_uart_tx_sink[_named] now honor echo_stdout for the twin (silence only —
  which instance is a console is decided at construction), so capture-only
  callers stop getting a duplicate host-console echo.
- browser_c3_fast_start_wifi_associates stopped the instant the body arrived,
  i.e. exactly where the device wedges, so it passed against a broken device.
  It now runs through PARSED to the repaint and asserts the full body arrives.

Reset oracle unaffected: no UART register is in the asserted set (the C3 console
regs are excluded as ROM-touched). Walk participation unchanged — the twin is
uses_scheduler() + matrix level export, so the OLED rom-boot bus stays
walk-deletable.
