+++ crates/core/src/peripherals/generic_factory.rs
@@ -99,6 +99,10 @@ pub const MODEL_TYPES: &[&str] = &[
     "esp32c3_apb_saradc",
     "esp32c3_ledc",
     "esp32c3_rmt",
+    // MUST be listed: the fuzzy fallback matches `contains("uart")`, which
+    // would coerce the C3's UART onto the STM32 register map — the silently
+    // wrong model that wedged every `Serial.print` over 128 bytes.
+    "esp32c3_uart",
     // nRF52 behavioral models (nrf52 factory).
     "nrf52840_twim",
     "nrf52_saadc",
