+++ validation/manifest.yaml
@@ -600,7 +600,31 @@ boards:
     # stack). Reset-oracle registers and layout untouched; esp32c3_walk_differential
     # (incl. framebuffer @ tick 512) and reset conformance remain the offline gates.
     # No live USB-JTAG re-capture this session.
-    drift_ack: 2026-07-27
+    # 2026-07-28: UART0/UART1 now use the real Espressif register map. The chip
+    # yaml declared them as the vendor-neutral `type: "uart"`, which resolves to
+    # the STM32 layout (SR@0x00, DR@0x04) — so the C3's FIFO writes landed on a
+    # status register, STATUS@0x1C read back 0 (uart_ll_get_txfifo_len() saw all
+    # 128 entries free, always), and NOTHING ever asserted the UART interrupt
+    # (the 2026-07-18 note above already recorded "the ESP32-C3 ... declares
+    # uart0 with NO irq:"). ESP-IDF's blocking `uart_tx_all()` fills the TX FIFO
+    # and, for any write longer than it, enables UART_INTR_TXFIFO_EMPTY and
+    # blocks on tx_fifo_sem — which was never given, so any `Serial.print` over
+    # 128 bytes wedged the sketch forever (every task blocked, idle task in wfi).
+    # Reproduced on the shipped stats lab: the 165-byte HTTP body truncated at
+    # exactly 128 chars and the e-paper panel never left "FETCHING STATS".
+    # Both UARTs now wire crates/core/src/peripherals/esp32c3/uart.rs, which
+    # binds the shared Espressif twin (relocated esp32s3/uart.rs ->
+    # peripherals/esp_uart.rs; same IP, chip-neutral now that two families use
+    # it) with the C3's matrix sources (UART0=21, UART1=22) and 160 MHz core
+    # clock. Reset-oracle UNAFFECTED: no UART register is in the asserted set —
+    # the C3 UART console regs are explicitly excluded as ROM-touched/dynamic
+    # (see esp32c3_reset_conformance module docs). Walk participation unchanged:
+    # the twin is uses_scheduler() + matrix_irq_sources level export, so the
+    # OLED rom-boot bus stays walk-deletable (EXPECTED_PINNERS still empty).
+    # Serial RX injection keeps parity — the twin now exposes rx_buffer() and
+    # attach_uart_rx_source[_named] resolve it, so `uart_injections:` still
+    # works. No live USB-JTAG re-capture this session.
+    drift_ack: 2026-07-28
 
   - id: nucleo-l476rg
     doc: docs/boards/nucleo-l476rg.md
